tracker/requirements.txt
aiofiles==24.1.0
asyncpg==0.30.0
aiosqlite==0.20.0
fastapi==0.115.6
fastjsonschema==2.21.1
httpcore==1.0.7
httpx==0.27.2
isort==5.13.2
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
pylint==3.3.3
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
setuptools==75.8.0
urllib3==2.3.0
uvicorn==0.34.0
validators==0.34.0
wheel==0.45.1
SQLAlchemy==2.0.37
pydantic[email]
